Fix missing data in ptw traces and high memory usage
In-flight threads were being silently lost at the end of program
execution, creating a wait group and waiting on it before terminating
the parser avoids this issue.

Also adds a second notifier for writer backpressure on parser to slow
the spawning of new threads. Possibly not needed, and should be
optimised.